# Optimize raw data job queries.