<!--- The title text will be used to populate Changelog and associated documentation.
      Please make sure that the title is a complete sentence -->

## Description
<!--- Describe your changes in detail -->
<!--- Include screenshots for GUI changes if appropriate -->
<!--- If any documentation outside of this repo needs to be added or updated,
      please include links to the relevant docs and how they should be changed -->
This PR optimizes raw data queries in the Jobs realm by using `WHERE` conditions on the `end_day_id` field of `modw.job_tasks` instead of the `end_time_ts` field. This PR also removes the `DISTINCT` modifier for single-day queries.

There is also a PR for https://github.com/ubccr/xdmod-xsede/pull/433.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

The new query is much faster (<10 seconds per single-day query compared to ~200 seconds), and in conjunction with https://github.com/ubccr/xdmod/pull/1792 and https://github.com/ubccr/xdmod-data/pull/19, it makes the data analytics framework more responsive.

## Tests performed
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
In addition to ensuring the regression tests from https://github.com/ubccr/xdmod/pull/1788 still pass:
1. Run the script from https://github.com/ubccr/xdmod-xsede/pull/436 with the `--time` argument.
    1. Set `'new': 'https://xdmod-dev.ccr.xdmod.org'` since https://github.com/ubccr/xdmod-xsede/pull/433 is already merged.
    1. Set `'old': 'https://xdmod-dev.ccr.xdmod.org:9001'` with the version of `classes/DataWarehouse/Query/Jobs/JobDataset.php` from https://github.com/ubccr/xdmod/pull/1779.
1. Run the script again with the `--time` argument and the variable `NUM_DAYS_PER_REQUEST` set to 1:
    1. Confirm the new Jobs realm queries take <10 seconds compared to ~200 seconds for the old queries.

## Checklist:
<!--- Go over all the following points and make sure they have all been completed -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] The pull request description is suitable for a Changelog entry
- [x] The milestone is set correctly on the pull request
- [x] The appropriate labels have been added to the pull request